ifelse function but then based on whether value is NULL or not

ifelse function but then based on whether value is NULL or not

Usage

ifelse0(value = NULL, alternative = NULL, allow_null = FALSE)

Arguments

value

metadata list object

alternative

alternative value

allow_null

can the alternative be NULL?

Value

value if non-NULL; alternative otherwise