Install default literature model

A very lightweight wrapper for model_from_api that installs previously published models packaged within PKPDsim.

Usage

install_default_literature_model(model, ...)

Arguments

model

Name of model, e.g., "pk_busulfan_mccune". See available_default_literature_models()

...

arguments passed onto model_from_api. For fine-grain control, it is better to install models directly from model_from_api() or new_ode_model().

Examples

if (FALSE) { # \dontrun{
install_default_literature_model("pk_busulfan_mccune")
} # }